requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
python-multipart==0.0.9

# ML Model dependencies
numpy==1.24.3
onnxruntime==1.18.0
torch==2.1.2; sys_platform != "darwin"
torch==2.1.2; sys_platform == "darwin" and platform_machine == "arm64"
torch==2.1.2; sys_platform == "darwin" and platform_machine == "x86_64"

# Data processing
pandas==2.1.4

# Security and auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9
limits==3.10.1

# Monitoring and metrics
prometheus-client==0.19.1

# Structured logging
python-json-logger==2.0.7

# Testing
pytest==8.2.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.27.0

# Development
black==24.4.2
flake8==7.0.0
mypy==1.10.0
